r6400: use the new locations for packages
If the package is referred for more than one time, the highest
version will be used. This selection might cause instable issues
later, which can be observed.

Generally, a pseudo-package like original layout, for example,
ap/gpl/timemachine to contain the dependent packages can be supported
with buildroot. currently, let's start with the single package
to speed the compilation.

Change-Id: I652bbb45cb7691e4f6302b3da4e77139fe2124dd